+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This is used for recording Invader's changes. This changelog is based on [Keep a
 - invader-bitmap: Height maps now generate bump maps similar to tool.exe
 - invader-bitmap: Added --usage (default or bumpmap) and --bump-height parameters
 - invader-bitmap: Added p8-bump support based on Stubbs the Zombie's palette
+- invader-bitmap: Dithering is now available for 16-bit and palettized bitmaps
 
 ### Fixed
 - invader-bitmap: Fixed some issues with spacing